// src/serverApi/v3/models/task-card-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HPI Schul-Cloud Server API
 * This is v3 of HPI Schul-Cloud Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { CardElementResponse } from './card-element-response';
import { TaskResponse } from './task-response';

/**
 * 
 * @export
 * @interface TaskCardResponse
 */
export interface TaskCardResponse {
    /**
     * The id of the task card
     * @type {string}
     * @memberof TaskCardResponse
     */
    id: string;
    /**
     * The title of the task card
     * @type {string}
     * @memberof TaskCardResponse
     */
    title: string;
    /**
     * Array of card elements
     * @type {Array<CardElementResponse>}
     * @memberof TaskCardResponse
     */
    cardElements?: Array<CardElementResponse>;
    /**
     * 
     * @type {string}
     * @memberof TaskCardResponse
     */
    courseName?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskCardResponse
     */
    courseId?: string;
    /**
     * Are the card elements draggable?
     * @type {boolean}
     * @memberof TaskCardResponse
     */
    draggable: boolean;
    /**
     * The task attached to the card
     * @type {TaskResponse}
     * @memberof TaskCardResponse
     */
    task: TaskResponse;
    /**
     * Visible at date of the task card
     * @type {string}
     * @memberof TaskCardResponse
     */
    visibleAtDate: string;
    /**
     * Due date of the task card
     * @type {string}
     * @memberof TaskCardResponse
     */
    dueDate: string;
}